Rename data-categories.slice --> taxonomy.slice
