+++ clients/ctl/admin-ui/src/features/taxonomy/index.ts
@@ -1,1 +1,1 @@
-export * from "./data-categories.slice";
+export * from "./taxonomy.slice";
